+++ Network.cs
@@ -20,14 +20,14 @@ namespace simple_network {
         //     //     layers[i] = new Layer(layerSizes[i],layerSizes[i+1]);
         //     // }
         //     layers[0] = new Layer(layerSizes[0],layerSizes[1]);
-        //     float[] b = {0.3f,-0.6f,0.5f};
+        //     double[] b = {0.3f,-0.6f,0.5f};
         //     layers[0].biases = b;
-        //     float[,] w = new float[2,3]
+        //     double[,] w = new double[2,3]
         //     w = {{0.0f,-0.3f},{-2.4f,-0.2f,0.7f}};
         //     layers[0].weights = w;
         // }
         
-        public static int PickClass(float[] inputs) {
+        public static int PickClass(double[] inputs) {
             int maxIndex = 0;
             for (int i = 0; i < inputs.Length; i++) {
                 if (inputs[i] > inputs[maxIndex]) maxIndex = i;
@@ -35,11 +35,11 @@ namespace simple_network {
             return maxIndex;
         }
 
-        public static float ActivationFunction(float input) {
-            return 1 / (1 + (float)Math.Exp(input));
+        public static double ActivationFunction(double input) {
+            return 1 / (1 + (double)Math.Exp(input));
         }
 
-        float[] GetOutputs(float[] inputs) {
+        double[] GetOutputs(double[] inputs) {
             foreach (Layer layer in layers) {
                 inputs = layer.GetLayerOutputs(inputs);
             }
@@ -47,8 +47,8 @@ namespace simple_network {
             return inputs;
         }
 
-        public int Classify(float[] inputs) {
-            float[] outputs = GetOutputs(inputs);
+        public int Classify(double[] inputs) {
+            double[] outputs = GetOutputs(inputs);
 
             return PickClass(outputs);
         }
